requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-RESTful==0.3.9
PyJWT==2.1.0
requests==2.26.0
cryptography==3.4.7
psutil==5.8.0
pymongo==4.3.3
psycopg2-binary==2.9.5
python-dotenv==0.19.0
gunicorn==20.1.0
gunicorn==20.1.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
python-dotenv==0.19.0